Migrated Cycle Node in Linked List
