+++ docs/linkedlist/CycleNodeInLinkedList.md
@@ -1,45 +1,22 @@
-<!----- Conversion time: 2.764 seconds.
-
-
-Using this Markdown file:
-
-1. Cut and paste this output into your source file.
-2. See the notes and action items below regarding this conversion run.
-3. Check the rendered output (headings, lists, code blocks, tables) for proper
-   formatting and use a linkchecker before you publish this page.
-
-Conversion notes:
-
-* Docs to Markdown version 1.0β14
-* Tue Feb 12 2019 15:45:57 GMT-0800 (PST)
-* Source doc: https://docs.google.com/open?id=1pahIjiU0tBxUIyU3owapmmHlQ1y6dsYdpUgQnsAFvqQ
-* This document has images: check for >>>>>  gd2md-html alert:  inline image link in generated source and store images to your server.
------>
-
-
-
 # Cycle Node in Linked List
 
 [https://app.laicode.io/app/problem/38](https://app.laicode.io/app/problem/38)
 
 [https://leetcode.com/problems/linked-list-cycle-ii](https://leetcode.com/problems/linked-list-cycle-ii)
 
-
 ## Description
 
 Check if a given linked list has a cycle. Return the node where the cycle starts. Return null if there is no cycle.
 
-__Medium__
+**Medium**
 
-__Linked List__
+**Linked List**
 
-__Two Pointers__
-
+**Two Pointers**
 
 ## Assumption
 
 The linked list should be a singly-linked, non-null linked list.
-
 
 ## Algorithm
 
@@ -48,36 +25,33 @@ The high level idea to tackle this problem is to
 1.  Make sure there is a cycle in the linked list
 2.  Look for the node where the cycle starts
 
-
 To implement a solution, we need to figure out the two steps separately.
 
 1.  Determine if there is a cycle in a given linked list
     1.  [Check if Linked List Has a Cycle](../../Easy/CheckIfLinkedListHasACycle)
-2.  Look for the starting node of the cycle
-    1.
+2.  Look for the starting node of the cycle 1.
 
-![alt_text](cyclenode.png "image_tooltip")
+    1.  To understand this solution, you just need to ask yourself these question.
 
-    2.  To understand this solution, you just need to ask yourself these question.
         1.  Assume the distance from head to the start of the loop is x1
         2.  the distance from the start of the loop to the point fast and slow meet is x2
         3.  the distance from the point fast and slow meet to the start of the loop is x3
         4.  What is the distance fast moved? What is the distance slow moved? And their relationship?
+
             1.  d(fast) = x1 + x2 + x3 + x2
             2.  d(slow) = x1 + x2
-            3.  d(fast) = 2 * d(slow) ⇒ x1 + x2 + x3 + x2 = 2 (x1 + x2)
+            3.  d(fast) = 2 \* d(slow) ⇒ x1 + x2 + x3 + x2 = 2 (x1 + x2)
             4.  Thus x1 = x3
 
                 Finally got the idea.
+
+![alt_text](CycleNodeInLinkedList.png "image_tooltip")
+
 **It is important that both slow and fast start moving from the head such that the distances they cover will always be calculated from the same point**
-
-
 
 ## Solution
 
-
 ### Code
-
 
 ```java
 /**
@@ -91,6 +65,7 @@ To implement a solution, we need to figure out the two steps separately.
  * }
  */
 public class Solution {
+
   public ListNode cycleNode(ListNode head) {
     // write your solution here
     // Step 1: determine if there is a cycle in the linked list
@@ -130,10 +105,7 @@ public class Solution {
 }
 ```
 
-
-
 ### Complexity
-
 
 #### Time
 
@@ -143,10 +115,6 @@ Step 2 takes O(n) to look for the starting node of the cycle.
 
 Therefore, O(n).
 
-
 #### Space
 
 Only ListNode objects with constant spaces were created ⇒ O(1)
-
-
-<!-- Docs to Markdown version 1.0β14 -->
